requirements.txt
discord.py==1.7.3
mysql_connector==0.3.1
requests==2.27.1
